rename folder types and files
